refactor(tauri): extract the context menus into their own module
Unit 11b. Moves the native context-menu cluster out of `lib.rs` into
`src-tauri/src/context_menu.rs` -- the implementation from the #923 banner
through `forward_context_menu_event`, plus its three `#[cfg(test)] mod`
blocks, which move with it so Rust privacy does not force broad APIs.

Verified verbatim: of the 567 non-blank lines moved, exactly seven differ,
and all seven are the `pub(crate)` prefix. The four items `lib.rs` still
calls need it; their three request structs need it for one reason and no
other, which the build found rather than a reviewer -- `#[tauri::command]`
refuses to generate a wrapper whose argument type is less visible than the
command. `Submenu` left `lib.rs`'s imports with the code that used it.

`#[tauri::command]` registers a command under its bare identifier, never a
module path, so the client's `invoke("show_context_menu", ...)` and the
`"context-menu-action"` event are untouched. There are no `#[cfg]` guards
in the span; the two OS branches are runtime matches on
`std::env::consts::OS`, so no CI leg's coverage changes.

This is the one cluster in the Unit 11 epic with no external guard test --
nothing in the repo reads `lib.rs` as text for a context-menu construct.
So the move takes one thing with it that colocation had been providing for
free: the `ctx:` prefix separating these ids from the tray's `MENU_*` ids
was convention, visible to anyone scrolling one file, asserted by nothing.
`context_menu_id_space_tests` pins both directions -- a context-menu id
without the prefix is silently dropped by the forwarding handler, and a
tray id with it is silently forwarded into the webview. The id set is
derived by running every builder over its whole (all-bool) input space
rather than listed, and `the_derivation_actually_finds_ids` is the control
on that sweep, since zero ids satisfies both assertions.

Also drops a false `#[allow(dead_code)]` on `has_selection`, which claimed
the field was reserved for future macOS use while four live call sites
read it.
